Make Observable an Abstract Class
Providing a default mutator for an Observable doesn't necessarily make
a lot of sense. Ideally Observable should be an interface, but JavaScript
does not provide syntax for that. Instead, an Abstract Class that forces
the implementation of a mutator to the SubClass makes a bit more sense
and makes the Observable a more flexible module.