Implement calculator operations
For the visual screen updates to make sense to the user, a "sub-screen"
presenting the last calculator state is needed together with an indicator
of any pending operation. This visual change also introduces the need to
modify the Calculator's _getState to return the additional necessary
information.

An input overwrite flag was deemed necessary to satisfy user expectations
when starting to type after getting a result, while at the same time
allowing chaining operations.